+++ src/Model/Table/PagesCategoriesTable.php
@@ -95,15 +95,7 @@ class PagesCategoriesTable extends AppTable
 
         $this->addBehavior('Timestamp');
         $this->addBehavior('MeCms.Tree');
-    }
 
-    /**
-     * Default validation rules
-     * @param \Cake\Validation\Validator $validator Validator instance
-     * @return \MeCms\Model\Validation\PostsCategoryValidator
-     */
-    public function validationDefault(\Cake\Validation\Validator $validator)
-    {
-        return new \MeCms\Model\Validation\PagesCategoryValidator;
+        $this->_validatorClass = '\MeCms\Model\Validation\PagesCategoryValidator';
     }
 }