+++ src/Model/Table/PhotosAlbumsTable.php
@@ -124,15 +124,7 @@ class PhotosAlbumsTable extends AppTable
         ]);
 
         $this->addBehavior('Timestamp');
-    }
 
-    /**
-     * Default validation rules
-     * @param \Cake\Validation\Validator $validator Validator instance
-     * @return \MeCms\Model\Validation\PhotosAlbumValidator
-     */
-    public function validationDefault(\Cake\Validation\Validator $validator)
-    {
-        return new \MeCms\Model\Validation\PhotosAlbumValidator;
+        $this->_validatorClass = '\MeCms\Model\Validation\PhotosAlbumValidator';
     }
 }